Server: Add CLI option to specify where to look for tracks
This way, the jar and configuration can be stored in separate
file hierarchies. This is useful when you have multiple track
directories, e.g. a small one with a selection of tracks and a larger
one.

It's also useful for starting a server with ssh without
having to pollute your server's home directory (`ssh user@host java
-jar server.jar --tracks-dir=/my/path/tracks`)