Server: Move tracks to resources, use them instead of local folder
With this change, the server no longer needs to be deployed alongside a
tracks directory and the symlink to the tracks directory can also be
removed, making the server work by default on Windows on IDEs.

Tracks are now bundled into the jar file generated by `mvn package`,
however the location where tracks are searched for can still be
overridden with a CLI flag.

This required changing the way FileSystemStatsManager and
FileSystemTrackManager load tracks by making it possible to configure
which kind of file system they should use as jar resources can only be
read as files when using a custom FileSystem.