Deobfuscate TextManager
This class works in a similar way to SoundManager, it contains a map for
localization strings that are specific to Minigolf as well as a map for
strings that were shared between multiple Playforia games.

As part of this I renamed Class89 -> LocalizationNode which represents
an XML node containing translations for a single entry.