Rename get_upd -> add in SynchronizedInteger
This more accurately reflects what the code does. The return value was
never used so I also changed the method to return void. Since get_upd()
without any parameters was just adding 1 to the internal integer, I
removed the method to simplify the class; now when 1 needs to be added
to the internal int, add(1) can be called.