Add CLI flag to set username
This makes automating set-up for testing a bit easier as it skips the
need to input a username.

To test for username vailidity, I reused the logic from
TrackTestLoginPanel and moved the validation logic into a separate
class.

This also removes the System.exit calls from Launcher.call; those code
paths were impossible to test since the whole Java instance was getting
terminated. By returning an integer instead, we can check the return
codes in tests.