Fix layout shift when resizing the game while in a lobby
Resizing the applet caused the lobby screen to do a new layout and
since it was using the default LayoutManager (FlowLayout), the panels
were being placed in unintended places. Most other classes explicitly
set the layout manager to null to avoid this, however, looks like this
was overlooked in LobbyPanel.