Fix client connection retry behaviour
Previously, this tried to open a new page in the browser, however,
in our environment, appletContext was null and was throwing an
exception.

This was fixed by destroying the game applet and reinitializing it when
the retry button is clicked.

Reinitalizing the applet caused its size to get re-measured which was
leading to layout bugs after retries, this was fixed by hardcoding the
applet width and height in AApplet. Since the whole game is hardcoded
for this size, this does not result in any problems elsewhere.