Paint app content in the center of the applet
Now, resizing the applet actually does something as the content is
painted in the middle of it. This was fairly easy to achieve since
almost all content is painted in the ContentPanel; it just had to be
centered in the applet by using a GridBagLayout. Other things directly
painted in the applet are the error texts and loading panel, which
needed their positioning to be adjusted.