Server: Sort TrackSets when they are created
This way, championship games run in the intended order, previously the
order tracks were read in was nondeterministic and the games could start
with the hardest track first.
Also, sort the list of TrackSets. This way the same message is sent from
the server to clients every time which makes testing easier.