Server: Fix single player quick start
Most of the code was already in place, the regular expression matching
packages in the server needed to be adjusted a bit. Also, the game code
assumed that everyone starting a game was already in a lobby so I
modified the quick start code to join the single player lobby before
launching a game.