Client: Remove unused TrackDumper code
Looks like this was originally used to fetch tracks and records from
the Playforia servers. It is no longer used in the codebase though
and can be removed.

TrackDumper contains a copy of the client connection code package
com.aapeli.connection which can all be removed as a better-debofuscated
copy is available in the client codebase.