Client: Remove Parameters's dependency on Applet parameters
Applet APIs are deprecated since Java 9. Instead of using an AppletStub
to pass parameters to the applet, we can just construct a map and pass
that to the Parameters class.

This initially broke retrying a connection in case of server restarts
since previously the Parameters were constructed every time in AApplet's
run method but now they're only constructed once upon start-up. The
retries were fixed fixed by removing the nulling of the params field in
AApplet's destroy field.